+++ src/cache/core/types/Cache.ts
@@ -28,7 +28,7 @@ export namespace Cache {
   }
 
   export interface EvictOptions {
-    id: string;
+    id?: string;
     fieldName?: string;
     args?: Record<string, any>;
     broadcast?: boolean;